requirements.txt
numpy
matplotlib
python-fcl
casadi
tqdm
pytest